Delete Veeam_365_KB4326.ps1
Veeam retracted this fix

from Veeam:

"After a joint investigation between Microsoft and Veeam®, it has been determined that the workaround may include undesired side-effects. As such, remediating actions are required as outlined below.
The following steps should be applied during the next available maintenance window in your Veeam Backup for Microsoft 365 environment:
1. Revert the product binaries to default:
The easiest way to accomplish this step is to reinstall the product or run an update. You can find the latest product builds here.
2. Perform a full sync of any affected backup addressed by the fix for the aforementioned error. To do so, from PowerShell, run:
Get-VBOJob -Name “Name of the affected job” | Start-VBOJob -RunAsync -Full 
It is important to take the above remediating steps as soon as possible. A cumulative patch is being developed by Veeam to change product behavior and avoid undesired consequences. The root cause of the issue has been resolved by Microsoft on the Microsoft 365 side."